requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
cloudscraper==1.2.64
PyGithub==1.56